Build approved packages and functions tibble

A utility function to help you build your approve package and functions list. This can be used by logrx to log unapproved use of packages and functions.

build_approved(pkg_list, file = NULL)

Arguments

pkg_list

A named list of character vectors where the name is the package name with a character vector of approved functions or 'All'

file

Name of file where the approved tibble will be read to. If not specified, the tibble is returned.

Default: NULL

Permitted Files: .RDS

Value

A tibble with two columns (library, function) and one row per function

Details

For more details see the help vignette:

vignette("approved", package = "logrx")

Examples

approved_pkgs <- list(
  base = c("library", "mean"),
  dplyr = "All"
)

# build and return
build_approved(approved_pkgs)
#> # A tibble: 290 × 2
#>    function_name library      
#>    <chr>         <chr>        
#>  1 library       package:base 
#>  2 mean          package:base 
#>  3 rows_upsert   package:dplyr
#>  4 src_local     package:dplyr
#>  5 db_analyze    package:dplyr
#>  6 n_groups      package:dplyr
#>  7 distinct      package:dplyr
#>  8 summarise_    package:dplyr
#>  9 group_split   package:dplyr
#> 10 group_by_if   package:dplyr
#> # … with 280 more rows

# build and save
dir <- tempdir()
build_approved(approved_pkgs, file.path(dir, "approved.rds"))