Formats and returns a vector of used package functions

write_used_functions() gets log.rx used_packages_functions attribute, formats and returns the list of used package functions

Usage

write_used_functions()

Value

Formatted vector of used package functions

Examples

if (FALSE) { # \dontrun{
write_used_functions()
} # }